requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
langchain
langchain-openai
chroma-hnswlib
chromadb
text-splitter
tiktoken
langchain-community
python-multipart
